+++ src/pages/onboarding/index.ts
@@ -1,0 +1,1 @@
+export { OnboardingPage } from './OnboardingPage';
